Making rest seperate
