+++ boilerplate_app/serializers.py
@@ -40,9 +40,3 @@ class ProjectsCreateSerializer(serializers.ModelSerializer):
     def create(self, validated_data):
         user = User.objects.get(pk=validated_data.pop('user'))
         return Projects.objects.create(**validated_data, user=user)
-
-
-class ProjectsListSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Projects
-        fields = ('id', 'project_name', 'user')
