+++ boilerplate_app/urls.py
@@ -12,6 +12,5 @@ urlpatterns = [
     path('login/', views.LoginView.as_view(), name='login-api'),
     path('logout/', views.LogoutView.as_view(), name='logout-api'),
     path('list/users/', views.UserAPIView.as_view(), name='user-api'),
-    path('project/', views.ProjectAPIView.as_view(), name='project-api'),
     path('docs/', schema_view.with_ui("swagger", cache_timeout=0), name="schema_view"),
 ]
