+++ boilerplate_app/views.py
@@ -9,10 +9,9 @@ from rest_framework.generics import GenericAPIView, CreateAPIView
 from rest_framework_jwt.serializers import JSONWebTokenSerializer
 
 from boilerplate_app.tasks import add
-from boilerplate_app.models import User, Projects
+from boilerplate_app.models import User
 from boilerplate_app.utils import generate_jwt_token
-from boilerplate_app.serializers import (
-    UserCreateSerializer, UserListSerializer, ProjectsCreateSerializer, ProjectsListSerializer)
+from boilerplate_app.serializers import UserCreateSerializer, UserListSerializer
 
 
 class TestAppAPIView(APIView):
@@ -122,53 +121,6 @@ class UserAPIView(GenericAPIView):
             return Response({'status': True,
                              'Response': users},
                             status=status.HTTP_200_OK)
-        except Exception as e:
-            return Response({'status': False, 'message': str(e)},
-                            status=status.HTTP_400_BAD_REQUEST)
-
-
-class ProjectAPIView(GenericAPIView):
-    serializer_class = ProjectsListSerializer
-    permission_classes = (IsAuthenticated,)
-
-    def get(self, request):
-        """
-        List all the projects.
-        """
-        try:
-            projects = Projects.objects.all()
-            project_serializer = ProjectsListSerializer(projects, many=True)
-
-            projects = project_serializer.data
-            return Response({'status': True,
-                             'Response': projects},
-                            status=status.HTTP_200_OK)
-        except Exception as e:
-            return Response({'status': False, 'message': str(e)},
-                            status=status.HTTP_400_BAD_REQUEST)
-
-    def post(self, request, format=None):
-        """
-        Create a project
-        """
-        try:
-            data = request.data
-            data['user'] = request.user.pk
-            serializer = ProjectsCreateSerializer(data=data)
-            if serializer.is_valid():
-                project = serializer.create(serializer.data)
-                return Response({'status': True,
-                                 'project': project.id,
-                                 'message': "Project Added Successfully"},
-                                status=status.HTTP_200_OK)
-            else:
-                message = ''
-                for error in serializer.errors.values():
-                    message += " "
-                    message += error[0]
-                return Response({'status': False,
-                                 'message': message},
-                                status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'status': False, 'message': str(e)},
                             status=status.HTTP_400_BAD_REQUEST)
