+++ rest/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path, include
+
+
+urlpatterns = [
+    path('v1/', include("rest.v1.urls")),
+]
