+++ rest/v1/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path, include
+
+
+urlpatterns = [
+    path('boilerplate-app/', include("rest.v1.boilerplate_app.urls")),
+]
